generate-machine-sdt: versal-vek280-sdt-seg: Fix default config to use SD1
vek280 does not have qspi, so hard code the flash type as OSPI.  If the
boot mode is changed to '8', the system will attempt to boot from the
OSPI instead of SD1.

Default is SD1 (2.0) emulation (mode 5).  Note the VEK280 user manual says
only SD1 (3.0) emulation (mode 14) is enabled on the actual hardware, but
QEMU does not support this mode, so use 2.0 instead.

This also fixes the ospi device to use 'mtd', not sd.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>