imgrcvry-vek385-system-conf.dtsi: Fix serial config
Image recovery on VEK280 was hanging due to shared device tree configuration with VEK385, which uses serial0 as the debug UART. This caused conflicts during the boot process.
To resolve the issue, a separate DTSI file has been created for VEK385.
This isolates the platform-specific serial configurations and ensures both VEK280 and VEK385 function independently and correctly during image recovery and normal boot.

Signed-off-by: Parthiban Kanchipuram <Parthiban.Kanchipuram@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>