meta-amd-adaptive-socs-bsp: qemu-microblaze-v-*.conf: Update machines
The 'qemu.inc' include will come from qemu-xilinx automatically, calling
it manually results in the QB_DEFAULT_FSTYPE always being set to ext4 which
is not desired.

The XILINX_WITH_ESW should be _cleared_ not set to "none".  This will ensure
that any special xsct or sdt processing is skipped.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Message-ID: <20250925214450.1170311-2-mark.hatle@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>